refactor: LandingPage파일 FSD에 맞게 리팩토링
