+++ src/features/LandingPageFilterTabs/lib/getLandingPageCategories.ts
@@ -1,0 +1,12 @@
+import type { CategoryCode } from '@/entities/post';
+import { axiosInstance } from '@/shared/api/base/axiosInstance';
+
+export interface Category {
+  code: CategoryCode;
+  displayName: string;
+}
+
+export async function getLandingPageCategories(): Promise<Category[]> {
+  const { data } = await axiosInstance.get<Category[]>('/posts/categories');
+  return data;
+}
