+++ src/features/LandingPageFilterTabs/model/useLandingPageCategoriesQuery.ts
@@ -1,0 +1,9 @@
+import { useQuery } from '@tanstack/react-query';
+import { getLandingPageCategories } from '../lib/getLandingPageCategories';
+
+export const useLandingPageCategoriesQuery = () => {
+  return useQuery({
+    queryKey: ['categories'],
+    queryFn: getLandingPageCategories,
+  });
+};
