+++ src/pages/landing/ui/LandingPage.tsx
@@ -1,6 +1,6 @@
 import { MorePostsButton } from '@/shared/ui';
-import { FilterTabs } from '@/widgets/FilterTabs';
-import { useFilter } from '@/widgets/FilterTabs/model/useFilter';
+import { LandingPageFilterTabs } from '@/features/LandingPageFilterTabs';
+import { useFilter } from '@/features/LandingPageFilterTabs/model/useFilter';
 import { HeroSection } from '@/widgets/HeroSection';
 import { PostList } from '@/widgets/PostList';
 import { SupportBand } from '@/widgets/SupportBand';
@@ -11,7 +11,7 @@ export function LandingPage() {
   return (
     <>
       <HeroSection />
-      <FilterTabs category={category} setCategory={setCategory} />
+      <LandingPageFilterTabs category={category} setCategory={setCategory} />
       <PostList className='mt-8' limit={6} code={category} showPagination={false} />
       <MorePostsButton to='/post' label='마음소식 더 보러가기' />
       <SupportBand />
