--- src/widgets/FilterTabs/index.ts
@@ -1,1 +1,0 @@
-export * from './ui/FilterTabs';
