+++ src/widgets/PostStats/index.ts
@@ -1,1 +1,1 @@
-export * from './ui/PostStats';
+export * from './ui/PostStars';
