+++ src/widgets/PostStats/ui/PostStars.tsx
